Change program to use the middle value of existing pixels in the histogram
I calculated the first and last position in the histogram that is not zero,
then get the middle position among them.
I did this because I wanted to use a better value for the threshold. It indeed
is better, but it can be improved calculating the half between the two highest hills

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>